fix(merge-tree): Apply various sided obliterate fixes (#23807)
## Description

Follow-up to #23780 which fixes more bugs in the implementation of sided
obliterate. Adds unit tests / minimized versions of the fuzz tests that
caught these issues.

High-level summary of the issues that were fixed:

- A visibility check in obliteration didn't correctly account for
clientId (in general, segment visibility requires a combination of
knowing the refSeq of the operation as well as which client performed
it; the latter part was missing)
- The logic around marking a segment as obliterated upon insertion was
subtly incorrect in a similar fashion (a segment inserted by some client
should never be affected by an obliterate from the same client)
- `SortedSegmentSet` did not correctly order local references when they
had multiple offsets on a given segment. Obliterates take a dependency
on this as they place local references on adjacent segments and use a
sorted segment set to query for overlapping bits.

With a combination of these fixes applied, an obliterate-focused
workload can now be enabled in the SharedString fuzz tests and I have
done so.

AFAIK, the remaining issues with obliterate are limited to:
- grouped batching (the obliterate farm in merge-tree added by @jzaffiro
simulates this kind of behavior using grouped ops)
- IntervalCollection
- Ops applied at varying refSeqs (the DDS fuzz harness does not do much
interesting here due to it always synchronizing all clients)

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>