ci: Add missing retries to dependency install steps (#23834)
## Description

Adds retries to a few dependency install steps that didn't have it. This
is a pattern we already follow in other places. It's especially relevant
because in the internal ADO project we use an intermediate feed that
mirrors npmjs and there can be some delay until that one sees public
packages.