refactor(ci): Centralized download feed (#23803)
## Description

Updates our pipelines to better leverage a centralized set of download
feeds, removing the need to create ad-hoc `.npmrc` files in different
places and making sure they're in the right location. The new
`$(ado-feeds-ff-download-only)` upstreams npmjs and all our internal
feeds, so _almost_ everything that any of our pipelines might need to
download when installing dependencies (including internal-only packages
like aria-logger and telemetry-generator) should be available there. The
exception is some internal packages in the `@microsoft` scope, which we
get from a separate feed. That should be a safe setup to use for
anything in our pipelines that only requires access to _download_
packages. The feed only gives read permissions to the build identity, so
it's not possible for pipeline runs to publish packages to this feed.

This fixes an issue where test pipelines could not see the
`fluid-framework` package published by a build pipeline, because .npmrc
only pointed certain scopes to our internal feeds, and that is an
unscoped package.

As part of this I realized that the `Npm@` tasks always override the
user-level `.npmrc` file they use (with one at
`/mnt/vss/_work/1/npm/<buildId>.npmrc`), so our approach of setting
`NPM_CONFIG_USERCONFIG` is not compatible with them. I replaced the ones
that were used for doing dependency installs with `Bash@3` tasks.

Of note: the use of the centralized `.npmrc` file also works for the
dependency installs done by our logic to install older versions of Fluid
for compat tests. For that case it was already hard to copy a
hand-crafted `.npmrc` file to the folder where `package.json` lives
because the folders where the installs happen are created dynamically
when the JavaScript code runs. That won't be a problem anymore.