feat(presence-tracker): Add "reactions" using presence notifications (#23294)
## Summary of changes

The presence-tracker example now uses the notifications infrastructure
from the presence package to send "reactions" to other clients. Users
can select any emoji from an emoji picker, and any left click in the
window will send the reaction to other users. Other clients will see the
reaction animate and fade out.

I also removed the previous "click to focus" box and replaced it with a
"cover" that shows/hides when focus is lost.

## Implementation notes

- Unlike the mouse and focus trackers, I did not create a class to
manage things. Instead, I just wrote an init function that initializes
the reactions for the whole app. I think this makes it a little easier
to see how things fit together, but could be refactored if needed.
- I considered a smaller/more limited reaction selection UX, but the
emoji picker was easy to adopt, so I went with that.