test: Remove --exit from mocha tests (#23804)
## Description

Removes the `--exit` flag explicitly added to npm scripts that run
`mocha`, where possible. Several packages still need it for their test
scripts to eventually terminate, so those stayed (see
https://github.com/microsoft/FluidFramework/pull/23804/commits/7cb47c7902bee273d5e592059939435327e2f849).

This flag usually indicates an issue in the tests (e.g. `setInterval()`
calls that were not cleared) that results in the process not shutting
down properly, and thus the npm script hanging. We should avoid it when
possible instead of making it a default in our `mocha` invocations.

For future reference: when running locally, the tests for
packages/test/snapshot are skipped unless the git submodule for
packages/test/snapshots/content has been downloaded ([per the repo's
README](https://github.com/microsoft/FluidFramework?tab=readme-ov-file#testing)).
They do run in CI, and that can cause discrepancies (like tests in PR
runs timing out, but not when run locally... until that submodule is
downloaded).