fix(server): Add necessary additionalContext in docker-compose.dev (#23860)
## Description

Currently, if you run `docker compose `-f docker-compose.dev.yml build`
from `./server`, the build will fail with errors like `failed to resolve
source metadata for docker.io/library/root:latest: pull access denied`.
This is because the docker-compose.dev.yml is missing
`additionalContext` for Historian and Gitrest to provide the correct
"root" context. This is present for all R11s services in the file, but
is missing for Historian and Gitrest.