fix(merge-tree): Rewrite partial lengths (#23859)
## Description

This fixes more bugs in obliterate fuzz testing, primarily by rewriting
partial lengths and cleaning up logic around when we force-recompute it.

Notes on prod changes:

- Overall, we are really inconsistent on where we try to update partial
lengths within merge-tree. This makes the incremental update codepath
difficult to maintain and it had a lot of wackiness as a result. I've
refactored a few codepaths in mergeTree to just explicitly force
recombination rather than attempt the incremental codepath, since I'm
not convinced that the old codepath did things right anyway.
[AB#31003](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31003)
tracks improving this area. In general, incremental updates are not
supported for overlapping removes/obliterates (any combination), as well
as when a segment is inserted + removed by another client all within the
collab window.
- Partial lengths was pretty much overhauled by reconceptualizing what
the fields mean. Notably, we can eliminate the concepts of
`remoteObliteratedLen` and all the overlapping fields by thinking about
what we previously called `cliSeqNumbers` (now `perClientAdjustments`)
as changes we need from the "base partial lengths" (i.e. entries for a
client with no changes) at a given refSeq. A lot of the time this does
align with ops submitted by that client, but not always. Overall I think
this makes the implementation quite a bit simpler and more possible to
reason about.

Notes on test changes:

- A couple tests tried to apply messages at the same seq from different
clients. This was caught by an assertion I added to partial lengths (in
the `addOrUpdate` codepath), so I fixed them.
- The obliterate farm tests required some tweaks to get a more
consistent model, but now mostly pass without `applyOpOnGeneration` (it
depends on the config, see notes there). About half of them pass with
`applyOpOnGeneration` now (before this change, it was 0). I suspect a
lot of the remaining failures may be related to an issue reproduced by
partial synchronization fuzz tests at a more rare rate (it stresses the
system less), which looks related to issues with insert tiebreaking
policy across merge blocks.
- One of the main changes here was making the operation runner support
having a client send multiple ops in succession, and having obliterate
not use group ops.
[AB#31001](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31001)
tracks deficiencies here.
- If we wanted, we could probably also run the obliterate farm with
applyOpOnGeneration with a bit less stressful of a config (the one in
main but moving opsPerRoundRange down to 20 passes almost all seeds,
with the only failure pointing to an annotate)

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>
Co-authored-by: Scarlett Lee <90647596+scarlettjlee@users.noreply.github.com>