Removed deprecated usages of the option summarizeProtocolTree [non-breaking] (#23355)
Removing deprecated code that usage of `summarizeProtocolTree` option
from the code. The change is non-breaking in nature. The PR only removes
the old unused usages of summarizeProtocolTree. (These changes enable
the breaking changes PR to be shorter and cleaner), and replaces it with
the config option `Fluid.Container.summarizeProtocolTree2` where needed
(i.e. the tests).

Context: `options.summarizeProtocolTree` in ILoaderOptions was used
initially to test the Single-commit summary feature with the host apps.
Now it is un-used option. Today, host apps even if they set
summarizeProtocolTree as false, ODSP driver will reject 2commit
summaries. So the option is misleading. The PR removed it's usage in the
code. It is expected that the `summarizeProtocolTree` decision is
provided by the driver layer through
`service.policies?.summarizeProtocolTree`.