Follow up to add nav parameter (#23845)
## Description

This task is a follow up to
[#20323](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/20323)
.

When loop tried to implement the code some errors came up, the expected
information was not being added.

Acceptance criteria
The link works properly when loop tries to implement and use it.

Execution plan
Update the methods we are using to receive the appName and the
containerPackageName. Also update IODSPResolvedURL to be able to have
the containerPackageName.

## Reviewer Guidance

Please let me know if there's something that is missing or that I should
keep in mind. Thanks!

Fixes:
[AB#29956](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/29956)

---------

Co-authored-by: Jatin Garg <48029724+jatgarg@users.noreply.github.com>