test(client-presence-tracker): fix still flaky test
This partially reverts commit 926ef8784fc099281295f42dd36d9f35b5e34bde.

Instead of waiting for count to be 1 when it is already expected to be 1, evaluate the attendee count and check on host side.

Oddly none of the other tests that have a timeout of 100ms waiting for attendee state appear to have complications.