Leverage 1ES pipeline templates schema for syntax checks in YAML files (#23862)
## Description

Enables using the schema for 1ES pipeline templates for Microsoft
contributors. The setting only works if signed in to the Azure Pipelines
extension with a `@microsoft.com` account, and does nothing otherwise.

Also removes the unused (and invalid according to the latest version of
the extension) `[azure-pipelines].customSchemaFile` setting.

The new setting removes syntax error warnings and provides IntelliSense
for 1ES-specific tasks when working on pipeline definition files.