tree: tests showcasing bug that hits assert 0x778 (#23921)
## Description

Tests that showcase a bug in tree view. When using a "complex schema"
(in this case an object, instead of just a primitive) for a tree, the
pattern of getting a view, initializing the tree through it, disposing
the view, getting a second view, and accessing the tree through it,
throws with assert 0x778. Any access to the tree root causes it.