Move Json domain from test to internal (#23043)
## Description

Export Json domain as internal. This is a step toward making it alpha,
but that will involve more tweaking of naming/scoping.

This makes it easy for users to embed JSON content in their trees.

This could be followed up with a similar domain for fluid serializable
content which is the same but allows handles: this will be needed for
map -> tree adapters.