Add FluidSerializableAsTree domain (#23927)
## Description

Adds a domain (schema and related types), currently internal, for
working with Add Fluid Serializable data.

This will be used by the schema for the map adapter to store that that
could have been put in a SharedMap value in
https://github.com/microsoft/FluidFramework/pull/23729

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

---------

Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>