(tree) Add changeset to include the performance enhancements in shared tree (#23926)
The op bunching feature
(https://github.com/microsoft/FluidFramework/pull/23732) improves the
performance of op processing in shared tree. This change adds a
changeset to include these performance enhancements that impacts the
users so that they are part of release notes.

Note that it repurposes an existing changeset added by
https://github.com/microsoft/FluidFramework/pull/23732. The changes in
that PR did not require a changeset but once was added incorrectly.