refactor(test-utils): Remove dependency on `aqueduct` (#23848)
Introduces internal copies of the needed types from `aqueduct`. I
stripped these copies down as much as possible given current test needs.
These can be cleaned up as needed in the future, as they are not
exported externally.

This is required prerequisite work for adding Tree-based data objects to
`aqueduct` (specifically, needed to add a dependency from `aqueduct` on
`tree`).