[release/2.22] fix(merge-tree): Correctly bookkeep insert into local-only obliterate (#23937)
## Description

Resolves some issues in obliterate's insert codepath related to unacked
obliterates over the range being inserted into. The gist of the change
is that we did not properly bookkeep a scenario where:

- The local client has an outstanding obliterate for a range
- Multiple other clients also obliterated this range and attempted to
insert into it

In this case, the local client knows that eventually, the remote clients
won't have won the rights to insert into that range (since the local
obliterate will remove the segment). However, remote clients won't know
this until the local client's obliterate is acked. Therefore if remote
clients make subsequent changes without seeing the local client's
obliterate, the local client may interpret them incorrectly.

The fix here is to fix the bookkeeping of the inserted segment on the
local client in the case above, which can be done by being a bit more
thorough on examining any overlapping obliterates (specifically,
tracking the newest acked obliterate and oldest unacked obliterate
separately from the newest overall). See unit tests for more details.


[AB#31009](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31009)

Cherry-pick of #23934 into 2.22 release branch.

Co-authored-by: Abram Sanderson <absander@microsoft.com>