Re-merge nav parameter (#23878)
## Description

There were some issues with the ODSP test tenants so we had to revert
the changes made in
[#29956](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/29956)
.

This task is meant to follow up on said revert to prepare the PR to
merge the changes again.

Execution Plan
Do the same changes as the PR in
[#29956](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/29956)

Acceptance Criteria
The changes are merged into the main repository of FF

## Reviewer Guidance

Let me know if there's anything that I should change or be aware about.

Fixes:
[AB#31180](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31180)

---------

Co-authored-by: Jatin Garg <48029724+jatgarg@users.noreply.github.com>