Passing the baseUrl to BasicRestWrapper (#23944)
## Description

There are cases in `BasicRestWrapper`, where we would pass the baseUrl
as `undefined` and instead pass the `baseUrl` + `apiEndpoint` in the
request args. However, when we log error telemetry, this approach might
cause the telemetry to indicate the request endpoint as "". This PR
fixes this bug. It should not have any impact on existing APIs.