refactor(aqueduct, fluid-static): Move `DataObjectKind` from `fluid-static` to `aqueduct` (#23799)
Moves `DataObjectKind` from `fluid-static` to `aqueduct`, and updates
`createDataObjectKind` (`aqueduct`) to use it.

This PR is a small part of a larger effort to work towards supporting
`SharedTree`-backed `DataObject`s (as opposed to `SharedDirectory`-based
ones).

Follow-up to #23784

### `presence` changes

Import changes in this PR resulted in more code being transitively
observed by the build in the `presence` package, which has
`exactOptionalPropertyTypes` configured. This was resulting in lib-check
failures in that package due to upstream packages violating
`exactOptionalPropertyTypes`.
This PR includes updates to the build of that package to mitigate these
issues until upstream packages have been sufficiently updated re:
`exactOptionalPropertyTypes`.

---------

Co-authored-by: Jason Hartman <jasonha@microsoft.com>