improvement(merge-tree): Remove wasMovedOnInsert (#23946)
## Description

With the changes in #23859, this field is now redundant with information
about when a given segment was inserted & when it was moved. This
removes anywhere that sets `wasMovedOnInsert`, and replaces areas that
read it with that computation.

For some peace of mind, I ran a build on the [first commit of this
PR](https://github.com/microsoft/FluidFramework/pull/23946/commits/2f6470174d8b433a7eb9e7e855d2b2d51ea5fdfa)
(which asserts that the old code path matches the new code path wherever
`wasMovedOnInsert` was read), and [it
passed](https://dev.azure.com/fluidframework/public/_build/results?buildId=324331&view=results).

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>