improvement(merge-tree): Avoid computing local partials when using local references (#23954)
## Description

Updates `MergeTree.nodeLength` to use the cached length in more cases.
Previously, we always passed `localSeq: undefined` when considering a
length perspective that was for the local client but included all local
changes. A refactor at some point changed a local reference codepath to
explicitly use the current localSeq on the collab window rather than
undefined, which causes some unnecessary computation in partial lengths.
This updates the logic in `nodeLength` to detect when a perspective
still includes all local changes and to just use the cached length if
so, which should give a small perf improvement for features that use
local references.
 

[AB#31013](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31013)

Co-authored-by: Abram Sanderson <absander@microsoft.com>