Move fail to core-utils (#23955)
## Description

Now that assert tagging can support `fail`, and that's confirmed to be
working in tree, it makes sense to broaden the availability of this
utility function.

## Breaking Changes

This will impact the call stack for asserts by adding "fail" to the top,
and the call stacks of "fail" by adding "assert" to the top.
Since such errors are mostly aggregated by their assert tag, and not
their stack, this shouldn't be a big deal, but it could have some
impact.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).