Fix cross version compatibility for op compression e2e tests (#23983)
## Description

**What:** [Pull Request
#18277](https://github.com/microsoft/FluidFramework/pull/18277) adds
cross version compatibility between the current version (2.X) and the
most recent major public release (1.X). Some op compression related
tests are failing the newly added compatibility tests and have been
disabled. We should find the root cause and fix in order to enable these
tests.

Affected file:
packages/test/test-end-to-end-tests/src/test/compression.spec.ts

**Acceptance Criteria:**
The root cause is identified and a fix or workaround is added to enable
these tests.
If applicable, we provide guidance to users on how to workaround this
issue when upgrading to 2.0

## Reviewer Guidance

The root cause is the fact that version 1.4.0 doesn't work well with
compression and grouping, so that was causing some test failures. Now
only that version is being skipped when cross compatibility is being
tested.

If a user is going to upgrade from 1.4.0 to 2.0, they have to be mindful
around the `ContainerRuntimeOptions` and make sure that if they have
grouping on, compression also has to be on.