refactor(container-runtime): Move summarizer-related code from containerRuntime.ts (#23990)
## Description

Moves code related to the summarizer from `containerRuntime.ts` to files
in `src/summary/`. Just trying to make `containerRuntime.ts` a more
manageable file.

All the moved code is types or methods/functions that don't depend on
the internals of the ContainerRuntime class. The `createSummarizer()`
function was inlined in its only call site inside
`formCreateSummarizerFn()`.