Optimizations for alias tests which are flaky and timeout (#23996)
There are a couple of aliasing end-to-end tests that are flaky and
intermittently time out. See
[AB#31828](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31828)
and
[AB#31831](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31831).
This PR attempts to fix that by optimizing a couple of steps and fixing
a bug:
1. In `beforeEach`, the test waits for container connection and calls
`provider.ensureSynchronized`. Both of these are not needed so removing
them.
2. In `Assign multiple data stores to the same alias, first write wins,
different containers`, a `provider.ensureSynchronized` is not needed so
removing it.
3. In `Assign multiple data stores to the same alias, first write wins,
different containers from snapshot` there is a bug where the test
expects the summarizer to summarize the data store that was created and
aliased. However, that may not happen because the
`provider.ensureSynchronized` is called before the summarizer is created
- it should be called after.
Also, the test object provider needs to be created with `syncSummarizer:
true` option for it to sync changes to the summarizer client.
4. Changed the awaits in the flaky tests to `timeoutAwait` so it's more
debuggable.