Add some (mostly failing) generic schema type tests (#23936)
## Description

This adds some tests which are generically parameterized over schema
types. These are inspired by ongoing experiments with a Table schema
generator, and issues encountered in its implementation.

Generally it seems that we have some cases where we would really like
TypeScript to simplify conditional type expressions based on extends
clauses, but it does not do so causing code which would otherwise type
check to fail. One such cause is covered in
https://github.com/microsoft/TypeScript/issues/52144#issuecomment-2686250788
but there are likely others.