Set debugAssert enablement to opposite of isInPerformanceTestingMode in tree benchmarks (#24049)
## Description

Set debugAssert enablement to opposite of isInPerformanceTestingMode in
tree benchmarks.

Note that debugAsserts are currently disabled by default, so this is
opting into them for correctness mode of perf tests.
This is designed to still work correctly if/when debug asserts are
enabled by default, and/or enabled for all tests.