Reduce calls from Deli to Historian (#24053)
## Description

After document inactivity or a fatal error, Deli will attempt to clean
up ECs by calling `deleteSummary` in Historian. In this PR there are two
changes to reduce the number of calls from Deli to Historian:

1) If the current time > the EC TTL + doc.createTime, then we can assume
that the container no longer exists on RedisFs and we need not invoke
`deleteSummary` explicitly.
2) If the current time <= the EC TTL + doc.createTime, then we ask
Historian to delete the EC summary. However, we let `requestWithRetry`
handle the logic to retry NetworkErrors. This will reduce calls to
Historian which continously result in a 404.

## Breaking Changes

The existing functionality of Deli does not differ. Hence, there are no
breaking changes.