test/staging-mode: Hold staged changes in PSM not Outbox (#24051)
Until now, the prototype has held all staged changes in a single batch
in the Outbox. This has required some awkward workarounds, indicating
it's not the most ergonomic way to model staged changes.

This PR switches it so that staged changes are allowed through to the
Pending State Manager, but marked so that they are never actually
submitted.

Upon exiting Staging Mode:
- If committing, the staged flag is removed from all pending state, and
(if we're connected) those changes are replayed. (if disconnected,
they'll be resubmitted upon connect)
- If discarding, we will rollback each pending message similar to how
Outbox's rollback works

This will provide the same functionality as the current prototype but in
a way that is much more ergonomic and maintainable.

---------

Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>