StagingMode: Move to experimental API (#24056)
Makes staging mode an experimental api. Overall, it feels a bit clunky
to use, I wonder if there is a better way to do it. I can think of some
other ways we could change how the new interfaces are exported, but it
wouldn't really change the bad things about the pattern