(tree) Docs, comments and other small updates to forest (#24017)
Updated a few comments and docs and renames a couple of files related to
forest. These are mostly irregularities I noticed when learning about
forest and related logic.
I renamed `schemaBasedEncoding.ts` to `schemaBasedEncode.ts` to keep it
consistent with the two other files in that folder which handle encoding
- `uncompressedEncode.ts` and `compressedEncode.ts`.