Remove deprecated summarizeProtocolTree option [breaking] (#24046)
Removed deprecated ILoaderOption = summarizeProtocolTree.

summarizeProtocolTree property in ILoaderOptions (the extended
definition in container layer) was added to test single-commit summaries
during the initial implementation phase. The flag is no longer required
or should be used, so marked it as deprecated. If a driver needs to
enable or disable single-commit summaries, it can do so via
IDocumentServicePolicies.


[AB#9098](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/9098)
#23320