improvement(merge-tree): Revamp in-memory representation to thoroughly leverage perspectives (#24047)
## Description

This PR updates:

- The in-memory representation of merge-tree segments to a more uniform
format
- Rather than use different property names for various different types
of operations and store them directly on segments, we use the notion of
an `OperationStamp` to indicate a particular point in time + client that
performed an operation. This allows a uniform representation for
provenance of an operation across different operation types (remove,
insert, obliterate)
- Removals and obliterates were previously tracked separately despite
the vast majority of code only caring about whether the segment was
visible or not from some given perspective. This led to significant
overcomplication of conditionals in many places, and in several places
that we don't have good coverage for (e.g. obliterate reconnection)
there were likely bugs where only one of the two types of removals was
handled. In the new format, removes and obliterates are tracked in the
same list.
- The API of merge-tree to more thoroughly leverage perspective changes
made in #24055 and this more uniform format

Besides general simplification of the APIs, this change helps toward
allowing clients to resolve other remote clients' positions at a
perspective other than 'all of their changes + refSeq' (for example
those sent via signals), as it introduces a reasonable API for tracking
this.

Specific non-goals of this PR:

- Change or introduce a new snapshot format
- Change any alpha+legacy merge-tree API surface
- Propagate API shape changes to sequence

Follow-up PRs can address some of these aspects.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>