DDS Fuzz: Make reducers sync for easier re-use (#24090)
The primary motivation for this change is to add support to the local
server stress harness for orderSequentially which will increase coverage
for batching and rollback. That will come as a follow up.
orderSequentially can only perform synchronous operations, and so I need
the models to be able to reduce synchronously, it is fine for generation
to remain async.

This also seems like a reasonable change as dds are generally
synchronous in nature. There are some dds like consensus which are not
synchronous, but those will be hard to test already given we can't track
asynchronous operations across fuzz operations, as operations much be
serializable, and the same techniques use to make that work, can make
them work with a synchronous reducer.