refactor(ci): Upload npm logs in test pipelines (#24094)
## Description

Adds a new yml template to upload npm (not pnpm!) logs as a tar file in
an Azure pipeline run. Updates the perf benchmarks and e2e tests
pipelines to leverage it. I only added it to these two for now because
those are the ones that run tests with compat for older versions and
thus cause `npm install` to be invoked by test-version-utils. I'm trying
to double check if npm is doing the default number of retries when it
fails to install packages during those installs, and I need the logs for
that. (With views towards potentially updating the config those installs
use, so transient issues while downloading packages during
compat-version installs don't cause hard failures in the pipeline).