Use encode safe characters in encodeCompactIdToString() (#24097)
`encodeCompactIdToString` utility function converts ids into a more
compact representation.

Earlier the function was mapping the ids to also use '{' and ']'
characters. This was problematic, because some old versions of driver
and runtime layer encode ids using `encodeURIComponent()`, which
transformed these characters to `%5B` and `%7B`, resulting in errors.
(One such issue is described in the PR description here
https://github.com/microsoft/FluidFramework/pull/24032)

This PR updates the function to only use encode safe characters which do
not change with `encodeURIComponent()`.


[AB#32688](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/32688)