Scrub non-ASCII characters from unexpectedAckReceived event (#24091)
This event is to help debug super rare cases where an op contents
changes during the roundtrip to the ordering service. We were logging
the first character where the JSON contents diverged, to learn some
about these cases. That's not a great practice for non-ASCII where a
single character can be a whole word.

We still want to see these characters if they're ASCII, e.g. to know is
it JSON structure, or a number, or a letter. But otherwise we don't need
to know anything beyond "not ASCII".