(test runtime mocks) Add ref seq# to messages in MockContainerRuntime during submit (#24102)
In the test runtime mocks, when a message is submitted in turn based
flush mode, the message is queued in the outbox until flush is
explicitly called. During flush, the message is assigned a reference
sequence number and processed. However, this behavior is not correct
because of two reason:
1. It doesn't preserve the correct state when the op was submitted. Say
the message was submitted when the last sequence number processed was
10. By the time flush is called, 10 more messages have been processed
making the last sequence number processed 20. The message will be
assinged reference seq# 20 which is not what the state was when the
message was submitted.
2. This is not inline with what the real container runtime does. It
assigns a reference sequence number during op submit even if it doesn't
submit it right away and queues it (say because it is not connected or
in read mode). This is the reference sequence number used when the op is
finally submitted.

This change adds reference seq# to the message when it's submitted and
then adds it to the outbox in turn based flush mode. When this message
is finally flushed and processed, this is the reference seq# used.

Note than if the client disconnects and reconnects between submit and
flush, the reference seq# will be regenerated - this part has not
changed.