Typing tweaks to reduce size of type declarations (#24088)
## Description

This introduces a new system type (which customers should never directly
reference) called "LeafSchema" which is used to reduce the verbosity of
our generated d.ts files, and likely simplify IntelliSense and type
errors as well.

Additionally JsonAsTree has had some unused imports added to it greatly
reduce the size of its generated d.ts file (avoid all the inline imports
all over).

This should hopefully make investigations of the non-deterministic
(based on incremental build) d.ts (and thus api report) generation for
the Iterator types within JsonAsTree.

Additionally, the ordering of types in JsonAsTree was adjusted to
hopefully make for easier to read output, and maybe if we get lucky
resolve the non-determinism (this is not expected or intended to be a
fix, but its possible it will fix it).