Local Server Stress: Support for orderSequentially and fix rollback bugs (#24105)
This change adds new a local server stress test which focuses on
orderSequentially and rollback. Rollback is a big part of the strategy
for our upcoming Staging Mode feature, and these tests will help us to
test rollback broadly and find and fix bug in parallel to feature
development.

In addition to adding the new test, this change also makes a number of
fixes to rollback in areas which were causing severe test issues.
- Added detection and throw error when changes are generated during
rollback
 - Ensure dirty state is accurate after rollback is executed
- Simply synchronization in local server stress and make it more
debugable
- Add fuzz feature where `"debug":true` can be added to an operation in
a fuzz replay file to launch the debugger when that operation is hit


[AB#32288](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/32288)

---------

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>
Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>